[skip ci] start by rearranging the files
