+++ uproot/write/expandingblock.py
