+++ uproot/write/objects/TFile.py
