+++ uproot/write/objects/beginkey.py
