+++ uproot/write/objects/directoryinfo.py
