+++ uproot/write/objects/headkey.py
