+++ uproot/write/objects/key.py
