+++ uproot/write/objects/streamerkey.py
